Upgrade iOS native SDK to 1.6.6.

+++ src/ios/StreetHawkCoreCompact.framework/Headers/PushDataForApplication.h
@@ -1,22 +1,22 @@
 /*
- * Copyright 2012 StreetHawk
+ * Copyright (c) StreetHawk, All rights reserved.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3.0 of the License, or (at your option) any later version.
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Created by Christine XYS.
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library.
  */
 
-#import "StreetHawkPublicFwds.h"
+#import <Foundation/Foundation.h>
+#import "SHTypes.h"
 
 /**
  Enum for actions.
@@ -149,7 +149,7 @@ typedef enum SHAppFGBG SHAppFGBG;
 /**
  Used for SHAction_OpenUrl to slide web page inside App. This indicates how many percentage screen should be covered by web page.
  */
-@property (nonatomic) CGFloat portion;
+@property (nonatomic) float portion;
 
 /**
  Used for SHAction_OpenUrl to slide web page inside App. This indicates the direction where web page slide in.
@@ -159,7 +159,7 @@ typedef enum SHAppFGBG SHAppFGBG;
 /**
  Used for SHAction_OpenUrl to slide web page inside App. This indicates how many seconds the animation takes.
  */
-@property (nonatomic) CGFloat speed;
+@property (nonatomic) float speed;
 
 /**
  The sound file name in notification payload. Normally no need to handle this in iOS, system play the sound automatically when notification arrives.
